# Perf refactor