--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/ValueCollectionsInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/ValueCollectionsInternal.cs
@@ -295,6 +295,7 @@ namespace Proto.Promises
                 }
             }
 
+            [MethodImpl(MethodImplOptions.NoInlining)]
             private void EnterCore()
             {
                 // Spin until we successfully get lock.