Refactored to improve performance.
Changed cancelations back to not suppress rejections.
Fixed a race condition with progress with a cancelation token.
Added `AllowStackToUnwind` property in the csproj for better stacktraces for debugging.

Master:

```
|         Type | Pending | Recursion |     Mean |   Error | Allocated | Survived |
|------------- |-------- |---------- |---------:|--------:|----------:|---------:|
|   AsyncAwait |    True |       100 | 278.4 μs | 4.70 μs |         - | 25,152 B |
| ContinueWith |    True |       100 | 464.1 μs | 4.32 μs |         - | 13,104 B |
```

This PR:

```
|         Type | Pending | Recursion |     Mean |   Error | Allocated | Survived |
|------------- |-------- |---------- |---------:|--------:|----------:|---------:|
|   AsyncAwait |    True |       100 | 191.9 μs | 3.24 μs |         - | 25,152 B |
| ContinueWith |    True |       100 | 403.5 μs | 3.92 μs |         - | 13,104 B |
```